+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GameValidationService.java
@@ -759,7 +759,7 @@ public class GameValidationService implements InitializingBean, PreferenceChange
     }
 
     //aliased ROM was found as nvram file
-    if (aliasedRom && (highscoreService.getSupportedRoms().contains(rom) || highscoreService.getSupportedRoms().contains(rom.toLowerCase()) || highscoreService.getSupportedRoms().contains(tableName))) {
+    if (aliasedRom && (highscoreService.isSupportedRom(rom) || highscoreService.isSupportedRom(rom.toLowerCase()) || highscoreService.isSupportedRom(tableName))) {
       validation.setRomIcon(GameScoreValidation.OK_ICON);
       validation.setRomIconColor(GameScoreValidation.OK_COLOR);
       validation.setRomStatus(GameScoreValidation.STATUS_ROM_ALIASED_MATCH_FOUND);
@@ -767,7 +767,7 @@ public class GameValidationService implements InitializingBean, PreferenceChange
     }
 
     //the ROM was found as nvram file
-    if (highscoreService.getSupportedRoms().contains(String.valueOf(rom)) || highscoreService.getSupportedRoms().contains(String.valueOf(rom).toLowerCase()) || highscoreService.getSupportedRoms().contains(tableName)) {
+    if (highscoreService.isSupportedRom(String.valueOf(rom)) || highscoreService.isSupportedRom(String.valueOf(rom).toLowerCase()) || highscoreService.isSupportedRom(tableName)) {
       validation.setRomIcon(GameScoreValidation.OK_ICON);
       validation.setRomIconColor(GameScoreValidation.OK_COLOR);
       validation.setRomStatus(GameScoreValidation.STATUS_ROM_MATCH_FOUND);
@@ -810,7 +810,7 @@ public class GameValidationService implements InitializingBean, PreferenceChange
     }
 
     //ROM is not supported
-    if (!StringUtils.isEmpty(rom) && HighscoreType.NVRam.equals(game.getHighscoreType()) && (highscoreService.getSupportedRoms().contains(rom) || highscoreService.getSupportedRoms().contains(rom.toLowerCase()))) {
+    if (!StringUtils.isEmpty(rom) && HighscoreType.NVRam.equals(game.getHighscoreType()) && (highscoreService.isSupportedRom(rom) || highscoreService.isSupportedRom(rom.toLowerCase()))) {
       validation.setValidScoreConfiguration(false);
       validation.setRomIcon(GameScoreValidation.ERROR_ICON);
       validation.setRomIconColor(GameScoreValidation.ERROR_COLOR);