+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/highscores/parsing/ScoreListAdapter.java
@@ -4,6 +4,7 @@ import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.highscores.Score;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
+import java.io.IOException;
 import java.util.Date;
 import java.util.List;
 
@@ -12,5 +13,5 @@ public interface ScoreListAdapter {
   boolean isApplicable(@NonNull Game game);
 
   @NonNull
-  List<Score> getScores(@NonNull Game game, @NonNull Date createdAt, @NonNull List<String> lines, boolean parseAll);
+  List<Score> getScores(@NonNull Game game, @NonNull Date createdAt, @NonNull List<String> lines, boolean parseAll) throws IOException;
 }