+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/highscores/parsing/nvram/NvRamParsingWithParser.java
@@ -7,14 +7,16 @@ import java.util.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import de.mephisto.vpin.commons.SystemInfo;
+import de.mephisto.vpin.restclient.system.ScoringDB;
 import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.highscores.Score;
 import de.mephisto.vpin.server.highscores.parsing.ScoreListAdapter;
-import edu.umd.cs.findbugs.annotations.NonNull;
 
 import net.nvrams.mapping.NVRamParser;
 import net.nvrams.mapping.NVRamScore;
 import net.nvrams.mapping.map.NVRamMapParser;
+import net.nvrams.mapping.pinemhi.NVRamPinemhiParser;
 import net.nvrams.mapping.superhac.NVRamSuperhacParser;
 
 /**
@@ -24,34 +26,30 @@ import net.nvrams.mapping.superhac.NVRamSuperhacParser;
  * - as a ScoreListAdapter, to parse raw to Scores
  */
 public class NvRamParsingWithParser implements NvRamOutputToRaw, ScoreListAdapter {
-  private final static Logger LOG = LoggerFactory.getLogger(NvRamOutputToRawWithPinemhi.class);
+  private final static Logger LOG = LoggerFactory.getLogger(NvRamParsingWithParser.class);
 
   // the parser which most of the work is delegated to
   private NVRamParser parser;
-  // The cached list of supported roms to accelerate checks
-  private Set<String> supportedRoms;
 
 
-  public static NvRamParsingWithParser createSuperhacParser() throws IOException {
-    NVRamParser superhacParser = new NVRamSuperhacParser();
+  public static NvRamParsingWithParser createPinemhiParser(ScoringDB scoringDB) throws IOException {
+    NVRamParser pinemhiParser = new NVRamPinemhiParser(SystemInfo.RESOURCES + "pinemhi/", 
+        scoringDB.getHighscoreTitles(), scoringDB.getHighscoreSkipTitlesCheck());
+    return new NvRamParsingWithParser(pinemhiParser);
+  }
+
+  public static NvRamParsingWithParser createSuperhacParser(ScoringDB scoringDB) throws IOException {
+    NVRamParser superhacParser = new NVRamSuperhacParser(SystemInfo.RESOURCES + "superhac/roms.json");
     return new NvRamParsingWithParser(superhacParser);
   }
 
-  public static NvRamParsingWithParser createNvramMapParser() throws IOException {
-    NVRamParser mapParser = new NVRamMapParser();
+  public static NvRamParsingWithParser createNvramMapParser(ScoringDB scoringDB) throws IOException {
+    NVRamParser mapParser = new NVRamMapParser(SystemInfo.RESOURCES + "maps/");
     return new NvRamParsingWithParser(mapParser);
   }
 
-  private NvRamParsingWithParser(NVRamParser parser) throws IOException {
+  NvRamParsingWithParser(NVRamParser parser) throws IOException {
     this.parser = parser;
-    this.supportedRoms = new HashSet<>();
-    for (String rom : parser.getSupportedNVRams()) {
-      supportedRoms.add(rom.toLowerCase());
-    }
-  }
-
-  public Set<String> getSupportedRoms() {
-    return supportedRoms;
   }
 
   @Override
@@ -59,32 +57,18 @@ public class NvRamParsingWithParser implements NvRamOutputToRaw, ScoreListAdapte
     return this.getClass().getSimpleName() + " [" + this.parser.getClass().getSimpleName() + "]";
   }
 
-  public boolean isSupportedRom(String rom) {
-    return supportedRoms.contains(rom.toLowerCase());
-  }
-
   //---------------------------------------
   // implementation of NvRamOutputToRaw
 
-  //@Override
-  public String convertOutputToRaw(@NonNull String nvRamFileName, File nvFile) throws Exception {
-    Locale locale = Locale.getDefault();
-    String rom = romFromNv(nvFile);
-    if (isSupportedRom(rom)) {
-      return String.join("\n", parser.getRaw(rom, nvFile, locale));
-    }
-    return null;
+  @Override
+  public boolean isSupportedRom(String rom) {
+    return parser.isSupportedRom(rom);
   }
 
-  public String romFromNv(File nvFile) throws IOException {
-    String rom = nvFile.getName();
-    int dotIndex = rom.lastIndexOf('.');
-    if (dotIndex > 0) rom = rom.substring(0, dotIndex);
-    int hyphenIndex = rom.indexOf('-');
-    if (hyphenIndex > 0) rom = rom.substring(0, hyphenIndex);
-    return rom;
+  @Override
+  public List<String> getRaw(String rom, File nvFile, Locale locale) throws IOException {
+    return parser.getRaw(rom, nvFile, locale);
   }
-
 
   //---------------------------------------
   // implementation of ScoreListAdapter
@@ -96,28 +80,23 @@ public class NvRamParsingWithParser implements NvRamOutputToRaw, ScoreListAdapte
   }
 
   @Override
-  public List<Score> getScores(Game game, Date createdAt, List<String> lines, boolean parseAll) {
+  public List<Score> getScores(Game game, Date createdAt, List<String> lines, boolean parseAll) throws IOException {
     Locale locale = Locale.getDefault();
     String rom = game != null ? game.getRom().toLowerCase() : "<no rom>";
     if (isSupportedRom(rom)) {
-      try {
-        List<NVRamScore> nvRamScores = parser.parseRaw(rom, lines, locale, parseAll);
-        List<Score> scores = new ArrayList<>();
-        for (NVRamScore nvramScore : nvRamScores) {
-          Score sc = new Score(createdAt, game.getId(),
-              nvramScore.getPlayerInitials(), null,
-              nvramScore.getRawScore(),
-              nvramScore.getScore(),
-              nvramScore.getPosition());
-          sc.setSuffix(nvramScore.getSuffix());
-          sc.setLabel(nvramScore.getLabel());
-          scores.add(sc);
-        }
-        return scores;
+      List<NVRamScore> nvRamScores = parser.parseRaw(rom, lines, locale, parseAll);
+      List<Score> scores = new ArrayList<>();
+      for (NVRamScore nvramScore : nvRamScores) {
+        Score sc = new Score(createdAt, game.getId(),
+            nvramScore.getInitials(), null,
+            nvramScore.getRawScore(),
+            nvramScore.getScore(),
+            nvramScore.getPosition());
+        sc.setSuffix(nvramScore.getSuffix());
+        sc.setLabel(nvramScore.getLabel());
+        scores.add(sc);
       }
-      catch (Exception ioe) {
-        LOG.error("Error while getting scores for game {}: {}", game, ioe.getMessage());
-      }
+      return scores;
     }
     return Collections.emptyList();
   }